refactor: decouple haptic playback and UI components by modularizing models, widgets, and logic into specific directories
