+++ example_app/lib/main.dart
@@ -1,179 +1,23 @@
-import 'dart:async';
-import 'dart:convert';
 import 'dart:io' show Directory, File, Platform;
-import 'dart:math' as math;
 import 'dart:typed_data';
 
 import 'package:audioplayers/audioplayers.dart';
 import 'package:file_picker/file_picker.dart';
-import 'package:flutter/foundation.dart' show ValueListenable, compute;
+import 'package:flutter/foundation.dart' show compute;
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart' show AssetManifest, rootBundle;
-import 'package:gaimon/gaimon.dart';
 import 'package:haptify/haptify.dart';
 
+import 'models/conversion_result.dart';
+import 'models/waveform_envelope.dart';
+import 'models/sample.dart';
+import 'playback/playback_controller.dart';
+import 'widgets/mode_selector.dart';
+import 'widgets/result_card.dart';
+import 'widgets/section_header.dart';
+import 'widgets/support_banner.dart';
+import 'widgets/visualizers.dart';
+
 void main() => runApp(const HaptifyDemoApp());
-
-/// A bundled sound with haptics pregenerated by `haptify convert` at build
-/// time (see `assets/haptics/`).
-class Sample {
-  const Sample({
-    required this.name,
-    required this.asset,
-    required this.ahap,
-    required this.timings,
-    required this.amplitudes,
-  });
-
-  final String name;
-  final String asset;
-  final String ahap;
-  final List<int> timings;
-  final List<int> amplitudes;
-}
-
-/// Builds the bundled-sample list from whatever lives in `assets/audio/`,
-/// pairing each sound with its pregenerated haptics in
-/// `assets/haptics/ahap/<name>.ahap` and
-/// `assets/haptics/waveform/<name>.haptic.json` (both emitted by
-/// `haptify convert`). Drop a new sound plus its converted haptics into
-/// those folders and it shows up here — no code changes. Sounds without
-/// matching haptics are skipped.
-Future<List<Sample>> loadBundledSamples() async {
-  final manifest = await AssetManifest.loadFromAssetBundle(rootBundle);
-  final assets = manifest.listAssets().toSet();
-  final audio = assets.where((k) {
-    return k.startsWith('assets/audio/') &&
-        (k.endsWith('.mp3') || k.endsWith('.wav'));
-  }).toList()..sort();
-
-  final result = <Sample>[];
-  for (final key in audio) {
-    final base = key.split('/').last.replaceAll(RegExp(r'\.(mp3|wav)$'), '');
-    final ahapKey = 'assets/haptics/ahap/$base.ahap';
-    final waveformKey = 'assets/haptics/waveform/$base.haptic.json';
-    if (!assets.contains(ahapKey) || !assets.contains(waveformKey)) continue;
-
-    final ahap = await rootBundle.loadString(ahapKey);
-    final waveform =
-        jsonDecode(await rootBundle.loadString(waveformKey))
-            as Map<String, dynamic>;
-    result.add(
-      Sample(
-        name: _prettifySampleName(base),
-        // audioplayers' AssetSource is rooted at `assets/`.
-        asset: key.substring('assets/'.length),
-        ahap: ahap,
-        timings: (waveform['timings'] as List).cast<int>(),
-        amplitudes: (waveform['amplitudes'] as List).cast<int>(),
-      ),
-    );
-  }
-  return result;
-}
-
-/// `spring_03` → `Spring 03`, `piano-loop` → `Piano Loop`.
-String _prettifySampleName(String base) => base
-    .split(RegExp(r'[_\-\s]+'))
-    .where((word) => word.isNotEmpty)
-    .map((word) => '${word[0].toUpperCase()}${word.substring(1)}')
-    .join(' ');
-
-/// Plays the whole authored haptic pattern in one shot with gaimon: AHAP on
-/// iOS, the haptify-rendered waveform on Android.
-void playHaptic({
-  required String ahap,
-  required List<int> timings,
-  required List<int> amplitudes,
-}) {
-  if (Platform.isIOS) {
-    Gaimon.patternFromData(ahap);
-  } else {
-    Gaimon.patternFromWaveForm(timings, amplitudes, false);
-  }
-}
-
-/// How the synced haptics are produced while a clip plays.
-enum HapticMode {
-  /// Hand the whole authored AHAP / waveform to Core Haptics / the vibrator
-  /// once — the smooth, high-fidelity signal haptify generated.
-  native,
-
-  /// Sample the intensity curve as the sound plays and fire discrete impacts —
-  /// always locked to playback, but buzzier.
-  pulsed,
-}
-
-/// The haptic intensity over time for one track, reconstructed from the
-/// Android waveform (segment durations + 0-255 amplitudes). Used both to drive
-/// continuous playback and to draw the visualizers.
-class HapticEnvelope {
-  HapticEnvelope(this.timings, this.amplitudes)
-    : durationMs = timings.fold<int>(0, (sum, ms) => sum + ms).toDouble();
-
-  final List<int> timings;
-  final List<int> amplitudes;
-  final double durationMs;
-
-  /// Intensity (0..1) of the segment active at [ms], or 0 past the end.
-  double intensityAt(double ms) {
-    if (ms < 0 || durationMs == 0) return 0;
-    var t = 0.0;
-    for (var i = 0; i < timings.length; i++) {
-      t += timings[i];
-      if (ms <= t) return amplitudes[i] / 255;
-    }
-    return 0;
-  }
-
-  /// A smooth loudness envelope resampled to [count] evenly spaced points —
-  /// the audio-side view of the same signal.
-  List<double> resample(int count) => [
-    for (var i = 0; i < count; i++) intensityAt((i + 0.5) / count * durationMs),
-  ];
-}
-
-/// The outcome of converting uploaded audio, shaped to cross the isolate
-/// boundary as plain data.
-class ConversionResult {
-  const ConversionResult({
-    required this.durationMs,
-    required this.transients,
-    required this.continuous,
-    required this.ahap,
-    required this.timings,
-    required this.amplitudes,
-    required this.warnings,
-    required this.pattern,
-  });
-
-  final int durationMs;
-  final int transients;
-  final int continuous;
-  final String ahap;
-  final List<int> timings;
-  final List<int> amplitudes;
-  final List<String> warnings;
-
-  /// The full [HapticPattern] for visualisation.
-  final HapticPattern pattern;
-}
-
-/// Runs haptify's full runtime pipeline off the UI thread.
-ConversionResult convertUploadedBytes(Uint8List bytes) {
-  final pattern = const AudioAnalyzer().analyzeBytes(bytes);
-  final waveform = pattern.toWaveform();
-  return ConversionResult(
-    durationMs: pattern.totalDuration.inMilliseconds,
-    transients: pattern.events.whereType<TransientEvent>().length,
-    continuous: pattern.events.whereType<ContinuousEvent>().length,
-    ahap: pattern.toAhap(),
-    timings: waveform.timings,
-    amplitudes: waveform.amplitudes,
-    warnings: [for (final w in waveform.warnings) w.message],
-    pattern: pattern,
-  );
-}
 
 class HaptifyDemoApp extends StatelessWidget {
   const HaptifyDemoApp({super.key});
@@ -196,36 +40,18 @@ class DemoPage extends StatefulWidget {
 }
 
 class _DemoPageState extends State<DemoPage> {
-  final AudioPlayer _player = AudioPlayer();
+  final PlaybackController _playback = PlaybackController();
 
-  // Playback / continuous-haptics driver.
-  final List<StreamSubscription<Object?>> _subs = [];
-  final ValueNotifier<double> _progress = ValueNotifier<double>(0);
-
-  /// A never-changing 0 for visualizers of tracks that aren't playing, so
-  /// their playhead sits at the start.
-  final ValueNotifier<double> _zeroProgress = ValueNotifier<double>(0);
-  final Stopwatch _sincePositionUpdate = Stopwatch();
-  Timer? _tick;
-  double _lastPositionMs = 0;
-  double _audioDurationMs = 0;
-  int _hapticTick = 0;
-  HapticMode _mode = HapticMode.native;
-
-  /// Bundled samples discovered under `assets/audio/`, loaded once at startup.
+  // Bundled samples discovered under assets/audio/, loaded once at startup.
   List<Sample> _samples = const <Sample>[];
   bool _loadingSamples = true;
 
-  /// Accordion state: the sample whose tile is currently open, plus a
-  /// controller per tile so opening one can collapse the previous.
+  // Accordion state: the open tile, plus a controller per tile so opening
+  // one collapses the previous.
   String? _expandedId;
   final Map<String, ExpansibleController> _tileControllers = {};
 
-  /// Identifies the track whose visualizers are currently live (a sample name
-  /// or `'upload'`), and its intensity envelope. Null when nothing plays.
-  String? _activeId;
-  HapticEnvelope? _activeEnv;
-
+  // Runtime-upload state.
   bool _converting = false;
   String? _uploadName;
   Uint8List? _uploadBytes;
@@ -236,29 +62,23 @@ class _DemoPageState extends State<DemoPage> {
   @override
   void initState() {
     super.initState();
-    _subs.add(
-      _player.onDurationChanged.listen((d) {
-        _audioDurationMs = d.inMilliseconds.toDouble();
-      }),
-    );
-    _subs.add(
-      _player.onPositionChanged.listen((p) {
-        _lastPositionMs = p.inMilliseconds.toDouble();
-        _sincePositionUpdate
-          ..reset()
-          ..start();
-      }),
-    );
-    // Only react to natural completion here: we drive stop() ourselves when a
-    // new track starts, and reacting to that stopped event would race the new
-    // track we set up right after.
-    _subs.add(
-      _player.onPlayerStateChanged.listen((s) {
-        if (s == PlayerState.completed) _stopPlayback();
-      }),
-    );
+    // Rebuild when the active track or haptic mode changes (button icons,
+    // selected segment). The playhead animates via its own ValueListenable.
+    _playback.addListener(_onPlaybackChanged);
     _loadSamples();
   }
+
+  @override
+  void dispose() {
+    _playback.removeListener(_onPlaybackChanged);
+    _playback.dispose();
+    for (final c in _tileControllers.values) {
+      c.dispose();
+    }
+    super.dispose();
+  }
+
+  void _onPlaybackChanged() => setState(() {});
 
   Future<void> _loadSamples() async {
     final loaded = await loadBundledSamples();
@@ -272,7 +92,6 @@ class _DemoPageState extends State<DemoPage> {
   /// Accordion behaviour: opening a tile collapses whichever one was open.
   void _onSampleExpanded(String name, bool expanded) {
     if (!expanded) {
-      // Ignore the collapse we trigger ourselves on the previous tile.
       if (_expandedId == name) _expandedId = null;
       return;
     }
@@ -283,171 +102,61 @@ class _DemoPageState extends State<DemoPage> {
     }
   }
 
-  @override
-  void dispose() {
-    _tick?.cancel();
-    for (final s in _subs) {
-      s.cancel();
-    }
-    _player.dispose();
-    _progress.dispose();
-    _zeroProgress.dispose();
-    for (final c in _tileControllers.values) {
-      c.dispose();
-    }
-    super.dispose();
-  }
-
-  /// The live playhead for the active track, or a static 0 for the rest.
-  ValueListenable<double> _progressFor(String id) =>
-      _activeId == id ? _progress : _zeroProgress;
-
-  /// Starts [playAudio] and, alongside it, a continuous stream of haptic
-  /// impacts that follow the pattern's intensity for as long as the sound
-  /// plays — so the phone buzzes throughout, not just at the onsets.
-  Future<void> _startPlayback({
-    required String id,
-    required HapticEnvelope env,
-    required Future<void> Function() playAudio,
-    required VoidCallback fireNativeHaptic,
-  }) async {
-    _tick?.cancel();
-    Gaimon.stop();
-    await _player.stop();
-    _lastPositionMs = 0;
-    _audioDurationMs = env.durationMs;
-    _hapticTick = 0;
-    _progress.value = 0;
-    _sincePositionUpdate
-      ..reset()
-      ..start();
-    setState(() {
-      _activeId = id;
-      _activeEnv = env;
-    });
-    // Native mode plays the whole authored pattern once; pulsed mode fires an
-    // impact per tick from _onTick instead. Either way the ticker below drives
-    // the shared playhead.
-    if (_mode == HapticMode.native) fireNativeHaptic();
-    await playAudio();
-    _tick = Timer.periodic(const Duration(milliseconds: 40), _onTick);
-  }
-
-  void _onTick(Timer timer) {
-    final env = _activeEnv;
-    if (env == null) {
-      _stopPlayback();
-      return;
-    }
-    final dur = _audioDurationMs > 0 ? _audioDurationMs : env.durationMs;
-    // Extrapolate between the coarse position events for a smooth playhead.
-    final est = (_lastPositionMs + _sincePositionUpdate.elapsedMilliseconds)
-        .clamp(0.0, dur)
-        .toDouble();
-    _progress.value = dur == 0 ? 0 : (est / dur).clamp(0.0, 1.0);
-
-    // In pulsed mode, fire an impact every other tick (~80ms) so it reads as
-    // continuous vibration rather than a single burst. Native mode already
-    // handed the whole pattern off at the start.
-    if (_mode == HapticMode.pulsed) {
-      _hapticTick++;
-      if (_hapticTick.isEven) _fireHaptic(env.intensityAt(est));
-    }
-
-    if (dur > 0 && est >= dur) _stopPlayback();
-  }
-
-  /// Maps live intensity to one of gaimon's impact strengths (all backed by
-  /// `HapticFeedback`, so they work on both platforms).
-  void _fireHaptic(double intensity) {
-    if (intensity < 0.12) return;
-    if (intensity < 0.4) {
-      Gaimon.light();
-    } else if (intensity < 0.7) {
-      Gaimon.medium();
-    } else {
-      Gaimon.heavy();
-    }
-  }
-
-  void _stopPlayback() {
-    _tick?.cancel();
-    _tick = null;
-    Gaimon.stop();
-    _player.stop();
-    _progress.value = 0;
-    if (_activeId != null) {
-      setState(() {
-        _activeId = null;
-        _activeEnv = null;
-      });
-    }
-  }
-
-  /// Plays the uploaded audio itself: straight from bytes on Android; via a
-  /// temp file on iOS, where audioplayers has no byte-source support.
-  Future<void> _playUploadSound() async {
-    final bytes = _uploadBytes;
-    if (bytes == null) return;
-    await _player.stop();
-    if (Platform.isAndroid) {
-      await _player.play(BytesSource(bytes));
-      return;
-    }
-    if (_uploadTempPath == null) {
-      final extension = _uploadName!.toLowerCase().endsWith('.wav')
-          ? 'wav'
-          : 'mp3';
-      final path = '${Directory.systemTemp.path}/haptify_upload.$extension';
-      await File(path).writeAsBytes(bytes, flush: true);
-      _uploadTempPath = path;
-    }
-    await _player.play(DeviceFileSource(_uploadTempPath!));
-  }
-
-  void _playUploadHaptic() {
-    final result = _result;
-    if (result == null) return;
-    // Haptics only: no audio, so the ticker's stopwatch drives the playhead.
-    // Routing through _startPlayback keeps it honouring the Native/Pulsed mode.
-    _startPlayback(
-      id: 'upload',
-      env: HapticEnvelope(result.timings, result.amplitudes),
-      playAudio: () async {},
-      fireNativeHaptic: () => playHaptic(
-        ahap: result.ahap,
-        timings: result.timings,
-        amplitudes: result.amplitudes,
-      ),
-    );
-  }
-
   void _playSample(Sample sample) {
-    _startPlayback(
+    _playback.start(
       id: sample.name,
-      env: HapticEnvelope(sample.timings, sample.amplitudes),
-      playAudio: () => _player.play(AssetSource(sample.asset)),
-      fireNativeHaptic: () => playHaptic(
-        ahap: sample.ahap,
-        timings: sample.timings,
-        amplitudes: sample.amplitudes,
-      ),
+      env: WaveformEnvelope(sample.timings, sample.amplitudes),
+      ahap: sample.ahap,
+      playAudio: (player) => player.play(AssetSource(sample.asset)),
     );
   }
 
   void _playUploadBoth() {
     final result = _result;
     if (result == null) return;
-    _startPlayback(
+    _playback.start(
       id: 'upload',
-      env: HapticEnvelope(result.timings, result.amplitudes),
-      playAudio: _playUploadSound,
-      fireNativeHaptic: () => playHaptic(
-        ahap: result.ahap,
-        timings: result.timings,
-        amplitudes: result.amplitudes,
-      ),
+      env: WaveformEnvelope(result.timings, result.amplitudes),
+      ahap: result.ahap,
+      playAudio: _playUploadSource,
     );
+  }
+
+  void _playUploadHaptic() {
+    final result = _result;
+    if (result == null) return;
+    _playback.start(
+      id: 'upload',
+      env: WaveformEnvelope(result.timings, result.amplitudes),
+      ahap: result.ahap,
+      playAudio: (_) async {},
+      awaitAudioStart: false,
+    );
+  }
+
+  Future<void> _playUploadSoundOnly() =>
+      _playback.playSoundOnly(_playUploadSource);
+
+  /// Plays the uploaded audio on [player]: straight from bytes on Android;
+  /// via a temp file on iOS, where audioplayers has no byte-source support.
+  Future<void> _playUploadSource(AudioPlayer player) async {
+    final bytes = _uploadBytes;
+    if (bytes == null) return;
+    if (Platform.isAndroid) {
+      await player.play(BytesSource(bytes));
+      return;
+    }
+    _uploadTempPath ??= await _writeUploadTemp(bytes);
+    await player.play(DeviceFileSource(_uploadTempPath!));
+  }
+
+  Future<String> _writeUploadTemp(Uint8List bytes) async {
+    final extension = (_uploadName ?? '').toLowerCase().endsWith('.wav')
+        ? 'wav'
+        : 'mp3';
+    final path = '${Directory.systemTemp.path}/haptify_upload.$extension';
+    await File(path).writeAsBytes(bytes, flush: true);
+    return path;
   }
 
   Future<void> _pickAndConvert() async {
@@ -485,12 +194,12 @@ class _DemoPageState extends State<DemoPage> {
       body: ListView(
         padding: const EdgeInsets.all(16),
         children: [
-          const _SupportBanner(),
-          _ModeSelector(
-            mode: _mode,
-            onChanged: (m) => setState(() => _mode = m),
+          const SupportBanner(),
+          ModeSelector(
+            mode: _playback.mode,
+            onChanged: (m) => _playback.mode = m,
           ),
-          const _SectionHeader(
+          const SectionHeader(
             'Bundled samples',
             'Every sound in `assets/audio/`, paired with the haptics '
                 '`haptify convert` generated for it. Tap to hear the sound and '
@@ -507,41 +216,9 @@ class _DemoPageState extends State<DemoPage> {
               child: Text('No sounds found in assets/audio/.'),
             )
           else
-            for (final sample in _samples)
-              Card(
-                clipBehavior: Clip.antiAlias,
-                child: ExpansionTile(
-                  key: PageStorageKey(sample.name),
-                  controller: _tileControllers.putIfAbsent(
-                    sample.name,
-                    ExpansibleController.new,
-                  ),
-                  onExpansionChanged: (expanded) =>
-                      _onSampleExpanded(sample.name, expanded),
-                  leading: IconButton(
-                    tooltip: _activeId == sample.name ? 'Stop' : 'Play',
-                    icon: Icon(
-                      _activeId == sample.name
-                          ? Icons.stop_circle
-                          : Icons.play_circle,
-                    ),
-                    onPressed: () => _activeId == sample.name
-                        ? _stopPlayback()
-                        : _playSample(sample),
-                  ),
-                  title: Text(sample.name),
-                  subtitle: Text('${sample.timings.length} waveform segments'),
-                  childrenPadding: const EdgeInsets.fromLTRB(16, 0, 16, 16),
-                  children: [
-                    _Comparison(
-                      env: HapticEnvelope(sample.timings, sample.amplitudes),
-                      progress: _progressFor(sample.name),
-                    ),
-                  ],
-                ),
-              ),
+            for (final sample in _samples) _sampleTile(sample),
           const SizedBox(height: 24),
-          const _SectionHeader(
+          const SectionHeader(
             'Convert your own',
             'Pick a WAV or MP3 from your device; haptify analyzes the bytes '
                 'at runtime — no files written, no ffmpeg.',
@@ -565,17 +242,18 @@ class _DemoPageState extends State<DemoPage> {
               ),
             ),
           if (_result case final result?)
-            _ResultCard(
+            ResultCard(
               name: _uploadName ?? 'upload',
               result: result,
-              isPlaying: _activeId == 'upload',
-              onToggle: () =>
-                  _activeId == 'upload' ? _stopPlayback() : _playUploadBoth(),
+              isPlaying: _playback.isPlaying('upload'),
+              onToggle: () => _playback.isPlaying('upload')
+                  ? _playback.stop()
+                  : _playUploadBoth(),
               onPlayHaptic: _playUploadHaptic,
-              onPlaySound: _playUploadSound,
-              visualizer: _Comparison(
-                env: HapticEnvelope(result.timings, result.amplitudes),
-                progress: _progressFor('upload'),
+              onPlaySound: _playUploadSoundOnly,
+              visualizer: Comparison(
+                env: WaveformEnvelope(result.timings, result.amplitudes),
+                progress: _playback.progressFor('upload'),
                 pattern: result.pattern,
               ),
             ),
@@ -583,400 +261,33 @@ class _DemoPageState extends State<DemoPage> {
       ),
     );
   }
-}
 
-/// Stacked "Audio" and "Haptics" visualizers sharing one moving playhead, so
-/// you can watch the haptic track track the sound it was generated from.
-class _Comparison extends StatelessWidget {
-  const _Comparison({required this.env, required this.progress, this.pattern});
-
-  final HapticEnvelope env;
-  final ValueListenable<double> progress;
-
-  /// When present (uploads), the haptic lane draws the real events —
-  /// continuous bars plus transient spikes. Bundled samples only carry a
-  /// waveform, so the lane falls back to segment blocks.
-  final HapticPattern? pattern;
-
-  @override
-  Widget build(BuildContext context) {
-    final scheme = Theme.of(context).colorScheme;
-    final audioSamples = env.resample(96);
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        _VizRow(
-          label: 'Audio',
-          icon: Icons.graphic_eq,
-          color: scheme.tertiary,
-          child: CustomPaint(
-            painter: _AudioPainter(audioSamples, progress, scheme.tertiary),
-            child: const SizedBox.expand(),
-          ),
-        ),
-        const SizedBox(height: 8),
-        _VizRow(
-          label: 'Haptics',
-          icon: Icons.vibration,
-          color: scheme.primary,
-          child: CustomPaint(
-            painter: _HapticPainter(
-              env,
-              progress,
-              scheme.primary,
-              pattern: pattern,
-            ),
-            child: const SizedBox.expand(),
-          ),
-        ),
-        const SizedBox(height: 6),
-        Text(
-          pattern != null
-              ? 'Same signal, two views: the audio loudness envelope, and the '
-                    'haptic events — continuous bars with transient spikes.'
-              : 'Same signal, two views: the audio loudness envelope and the '
-                    'waveform segments actually played as haptics.',
-          style: Theme.of(context).textTheme.bodySmall,
-        ),
-      ],
-    );
-  }
-}
-
-/// One labelled visualizer lane: an icon + title above a fixed-height canvas.
-class _VizRow extends StatelessWidget {
-  const _VizRow({
-    required this.label,
-    required this.icon,
-    required this.color,
-    required this.child,
-  });
-
-  final String label;
-  final IconData icon;
-  final Color color;
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            Icon(icon, size: 16, color: color),
-            const SizedBox(width: 6),
-            Text(
-              label,
-              style: Theme.of(context).textTheme.labelMedium?.copyWith(
-                color: color,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 4),
-        SizedBox(height: 44, width: double.infinity, child: child),
-      ],
-    );
-  }
-}
-
-/// Mirror-around-center bars — an oscilloscope-style view of the loudness the
-/// analyzer measured.
-class _AudioPainter extends CustomPainter {
-  _AudioPainter(this.samples, this.progress, this.color)
-    : super(repaint: progress);
-
-  final List<double> samples;
-  final ValueListenable<double> progress;
-  final Color color;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    if (samples.isEmpty) return;
-    final midY = size.height / 2;
-    final playX = progress.value * size.width;
-    final barW = size.width / samples.length;
-    final played = Paint()
-      ..color = color
-      ..strokeCap = StrokeCap.round
-      ..strokeWidth = math.max(barW * 0.6, 1.5);
-    final upcoming = Paint()
-      ..color = color.withValues(alpha: 0.28)
-      ..strokeCap = StrokeCap.round
-      ..strokeWidth = math.max(barW * 0.6, 1.5);
-
-    for (var i = 0; i < samples.length; i++) {
-      final x = (i + 0.5) * barW;
-      final h = samples[i] * size.height * 0.92;
-      canvas.drawLine(
-        Offset(x, midY - h / 2),
-        Offset(x, midY + h / 2),
-        x <= playX ? played : upcoming,
-      );
-    }
-    _drawPlayhead(canvas, size, playX, color);
-  }
-
-  @override
-  bool shouldRepaint(_AudioPainter old) =>
-      old.samples != samples || old.color != color;
-}
-
-/// The haptic timeline. With a [pattern] (uploads) it draws the real events —
-/// continuous bars plus transient spikes; otherwise it falls back to the
-/// Android waveform, one block per segment (width = duration, height =
-/// amplitude).
-class _HapticPainter extends CustomPainter {
-  _HapticPainter(this.env, this.progress, this.color, {this.pattern})
-    : super(repaint: progress);
-
-  final HapticEnvelope env;
-  final ValueListenable<double> progress;
-  final Color color;
-  final HapticPattern? pattern;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final playX = progress.value * size.width;
-    final p = pattern;
-    if (p != null) {
-      _paintEvents(canvas, size, p, playX);
-    } else {
-      _paintSegments(canvas, size, playX);
-    }
-    _drawPlayhead(canvas, size, playX, color);
-  }
-
-  void _paintEvents(Canvas canvas, Size size, HapticPattern p, double playX) {
-    final totalUs = p.totalDuration.inMicroseconds.toDouble();
-    if (totalUs == 0) return;
-    for (final event in p.events) {
-      final startX = event.time.inMicroseconds / totalUs * size.width;
-      final h = event.intensity * size.height;
-      final active = startX <= playX;
-      switch (event) {
-        case ContinuousEvent():
-          final endX = event.endTime.inMicroseconds / totalUs * size.width;
-          canvas.drawRRect(
-            RRect.fromRectAndCorners(
-              Rect.fromLTWH(startX, size.height - h, endX - startX, h),
-              topLeft: const Radius.circular(2),
-              topRight: const Radius.circular(2),
-            ),
-            Paint()..color = color.withValues(alpha: active ? 0.7 : 0.22),
-          );
-        case TransientEvent():
-          canvas.drawLine(
-            Offset(startX, size.height),
-            Offset(startX, size.height - h),
-            Paint()
-              ..color = color.withValues(alpha: active ? 1 : 0.3)
-              ..strokeWidth = 2.5
-              ..strokeCap = StrokeCap.round,
-          );
-      }
-    }
-  }
-
-  void _paintSegments(Canvas canvas, Size size, double playX) {
-    final dur = env.durationMs;
-    if (dur == 0) return;
-    final block = Paint()..color = color;
-    final faded = Paint()..color = color.withValues(alpha: 0.28);
-    var x = 0.0;
-    for (var i = 0; i < env.timings.length; i++) {
-      final w = env.timings[i] / dur * size.width;
-      final h = (env.amplitudes[i] / 255) * size.height;
-      final rect = RRect.fromRectAndCorners(
-        Rect.fromLTWH(x + 0.5, size.height - h, math.max(w - 1, 1), h),
-        topLeft: const Radius.circular(2),
-        topRight: const Radius.circular(2),
-      );
-      canvas.drawRRect(rect, x <= playX ? block : faded);
-      x += w;
-    }
-  }
-
-  @override
-  bool shouldRepaint(_HapticPainter old) =>
-      old.env != env || old.color != color || old.pattern != pattern;
-}
-
-void _drawPlayhead(Canvas canvas, Size size, double x, Color color) {
-  canvas.drawLine(
-    Offset(x, 0),
-    Offset(x, size.height),
-    Paint()
-      ..color = color
-      ..strokeWidth = 1.5,
-  );
-}
-
-/// Chooses how synced haptics are produced during playback: the whole authored
-/// pattern in one shot, or discrete impacts sampled from the intensity curve.
-class _ModeSelector extends StatelessWidget {
-  const _ModeSelector({required this.mode, required this.onChanged});
-
-  final HapticMode mode;
-  final ValueChanged<HapticMode> onChanged;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final caption = mode == HapticMode.native
-        ? 'Native: plays the whole Core Haptics / waveform pattern once — the '
-              'smooth, high-fidelity signal haptify generated.'
-        : 'Pulsed: fires discrete impacts sampled from the intensity curve as '
-              'the sound plays — always in sync, but buzzier.';
-    return Padding(
-      padding: const EdgeInsets.only(top: 8, bottom: 8),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          SegmentedButton<HapticMode>(
-            segments: const [
-              ButtonSegment(
-                value: HapticMode.native,
-                label: Text('Native'),
-                icon: Icon(Icons.graphic_eq),
-              ),
-              ButtonSegment(
-                value: HapticMode.pulsed,
-                label: Text('Pulsed'),
-                icon: Icon(Icons.blur_on),
-              ),
-            ],
-            selected: {mode},
-            onSelectionChanged: (s) => onChanged(s.first),
-          ),
-          const SizedBox(height: 6),
-          Text(caption, style: theme.textTheme.bodySmall),
-        ],
-      ),
-    );
-  }
-}
-
-class _SupportBanner extends StatelessWidget {
-  const _SupportBanner();
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder<bool>(
-      future: Gaimon.canSupportsHaptic,
-      builder: (context, snapshot) {
-        if (snapshot.data ?? true) return const SizedBox.shrink();
-        return Card(
-          color: Theme.of(context).colorScheme.errorContainer,
-          child: const Padding(
-            padding: EdgeInsets.all(16),
-            child: Text(
-              'This device reports no haptic support — simulators and some '
-              'emulators cannot vibrate. Run on a real phone to feel the '
-              'patterns.',
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
-
-class _SectionHeader extends StatelessWidget {
-  const _SectionHeader(this.title, this.body);
-
-  final String title;
-  final String body;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 12, top: 8),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(title, style: theme.textTheme.titleLarge),
-          const SizedBox(height: 4),
-          Text(body, style: theme.textTheme.bodySmall),
-        ],
-      ),
-    );
-  }
-}
-
-class _ResultCard extends StatelessWidget {
-  const _ResultCard({
-    required this.name,
-    required this.result,
-    required this.isPlaying,
-    required this.onToggle,
-    required this.onPlayHaptic,
-    required this.onPlaySound,
-    required this.visualizer,
-  });
-
-  final String name;
-  final ConversionResult result;
-  final bool isPlaying;
-  final VoidCallback onToggle;
-  final VoidCallback onPlayHaptic;
-  final VoidCallback onPlaySound;
-  final Widget visualizer;
-
-  @override
-  Widget build(BuildContext context) {
+  Widget _sampleTile(Sample sample) {
+    final playing = _playback.isPlaying(sample.name);
     return Card(
-      margin: const EdgeInsets.only(top: 12),
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(name, style: Theme.of(context).textTheme.titleMedium),
-            const SizedBox(height: 8),
-            Text(
-              '${result.durationMs} ms of haptics · '
-              '${result.transients} transients · '
-              '${result.continuous} continuous · '
-              '${result.timings.length} waveform segments',
-            ),
-            for (final warning in result.warnings)
-              Padding(
-                padding: const EdgeInsets.only(top: 4),
-                child: Text(
-                  warning,
-                  style: Theme.of(context).textTheme.bodySmall,
-                ),
-              ),
-            const SizedBox(height: 12),
-            Wrap(
-              spacing: 8,
-              runSpacing: 8,
-              children: [
-                FilledButton.icon(
-                  onPressed: onToggle,
-                  icon: Icon(isPlaying ? Icons.stop : Icons.play_arrow),
-                  label: Text(isPlaying ? 'Stop' : 'Sound + haptic'),
-                ),
-                FilledButton.tonalIcon(
-                  onPressed: onPlayHaptic,
-                  icon: const Icon(Icons.vibration),
-                  label: const Text('Haptic only'),
-                ),
-                FilledButton.tonalIcon(
-                  onPressed: onPlaySound,
-                  icon: const Icon(Icons.volume_up),
-                  label: const Text('Sound only'),
-                ),
-              ],
-            ),
-            const SizedBox(height: 16),
-            visualizer,
-          ],
+      clipBehavior: Clip.antiAlias,
+      child: ExpansionTile(
+        key: PageStorageKey(sample.name),
+        controller: _tileControllers.putIfAbsent(
+          sample.name,
+          ExpansibleController.new,
         ),
+        onExpansionChanged: (expanded) =>
+            _onSampleExpanded(sample.name, expanded),
+        leading: IconButton(
+          tooltip: playing ? 'Stop' : 'Play',
+          icon: Icon(playing ? Icons.stop_circle : Icons.play_circle),
+          onPressed: () => playing ? _playback.stop() : _playSample(sample),
+        ),
+        title: Text(sample.name),
+        subtitle: Text('${sample.timings.length} waveform segments'),
+        childrenPadding: const EdgeInsets.fromLTRB(16, 0, 16, 16),
+        children: [
+          Comparison(
+            env: WaveformEnvelope(sample.timings, sample.amplitudes),
+            progress: _playback.progressFor(sample.name),
+          ),
+        ],
       ),
     );
   }
