+++ example_app/lib/models/sample.dart
@@ -1,0 +1,68 @@
+import 'dart:convert';
+
+import 'package:flutter/services.dart' show AssetManifest, rootBundle;
+
+/// A bundled sound with haptics pregenerated by `haptify convert` at build
+/// time (see `assets/haptics/`).
+class Sample {
+  const Sample({
+    required this.name,
+    required this.asset,
+    required this.ahap,
+    required this.timings,
+    required this.amplitudes,
+  });
+
+  final String name;
+  final String asset;
+  final String ahap;
+  final List<int> timings;
+  final List<int> amplitudes;
+}
+
+/// Builds the bundled-sample list from whatever lives in `assets/audio/`,
+/// pairing each sound with its pregenerated haptics in
+/// `assets/haptics/ahap/<name>.ahap` and
+/// `assets/haptics/waveform/<name>.haptic.json` (both emitted by
+/// `haptify convert`). Drop a new sound plus its converted haptics into
+/// those folders and it shows up here — no code changes. Sounds without
+/// matching haptics are skipped.
+Future<List<Sample>> loadBundledSamples() async {
+  final manifest = await AssetManifest.loadFromAssetBundle(rootBundle);
+  final assets = manifest.listAssets().toSet();
+  final audio = assets.where((k) {
+    return k.startsWith('assets/audio/') &&
+        (k.endsWith('.mp3') || k.endsWith('.wav'));
+  }).toList()..sort();
+
+  final result = <Sample>[];
+  for (final key in audio) {
+    final base = key.split('/').last.replaceAll(RegExp(r'\.(mp3|wav)$'), '');
+    final ahapKey = 'assets/haptics/ahap/$base.ahap';
+    final waveformKey = 'assets/haptics/waveform/$base.haptic.json';
+    if (!assets.contains(ahapKey) || !assets.contains(waveformKey)) continue;
+
+    final ahap = await rootBundle.loadString(ahapKey);
+    final waveform =
+        jsonDecode(await rootBundle.loadString(waveformKey))
+            as Map<String, dynamic>;
+    result.add(
+      Sample(
+        name: _prettifySampleName(base),
+        // audioplayers' AssetSource is rooted at `assets/`.
+        asset: key.substring('assets/'.length),
+        ahap: ahap,
+        timings: (waveform['timings'] as List).cast<int>(),
+        amplitudes: (waveform['amplitudes'] as List).cast<int>(),
+      ),
+    );
+  }
+  return result;
+}
+
+/// `spring_03` → `Spring 03`, `piano-loop` → `Piano Loop`.
+String _prettifySampleName(String base) => base
+    .split(RegExp(r'[_\-\s]+'))
+    .where((word) => word.isNotEmpty)
+    .map((word) => '${word[0].toUpperCase()}${word.substring(1)}')
+    .join(' ');
