+++ example_app/lib/widgets/section_header.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+/// A titled section intro: a large title over a small body paragraph.
+class SectionHeader extends StatelessWidget {
+  const SectionHeader(this.title, this.body, {super.key});
+
+  final String title;
+  final String body;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 12, top: 8),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(title, style: theme.textTheme.titleLarge),
+          const SizedBox(height: 4),
+          Text(body, style: theme.textTheme.bodySmall),
+        ],
+      ),
+    );
+  }
+}
